requirements.txt
Flask==3.0.3
gunicorn==21.2.0
pyswisseph==2.10.3
pytz==2024.1
